requirements.txt
streamlit>=1.28.0
onnxruntime>=1.16.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
plotly>=5.17.0
streamlit-folium>=0.15.0
requests>=2.31.0
matplotlib>=3.7.0
seaborn>=0.12.0
ultralytics
openmeteo-requests
requests-cache
retry-requests
geopy
statsmodels